refactor-#156: change file imports for rest
