+++ src/microservices/customer/rest/controllers/customer.controller.ts
@@ -11,9 +11,8 @@ import {
 
 import { RestJwtAuthGuard } from '@auth/rest/guards';
 import { CustomerService } from '../../services';
-import { CreateCustomerProfileDto } from '../dto/create-customer-profile.dto';
-import { RestCustomer } from '../models/customer-rest.model';
-import { UpdateCustomerProfileDto } from '../dto/update-customer-profile.dto';
+import { CreateCustomerProfileDto, UpdateCustomerProfileDto } from '../dto';
+import { RestCustomer } from '../models';
 
 @Controller('customer')
 export class CustomerController {
