+++ src/microservices/customer/rest/dto/index.ts
@@ -1,0 +1,2 @@
+export * from './create-customer-profile.dto';
+export * from './update-customer-profile.dto';
