+++ src/microservices/customer/rest/models/index.ts
@@ -1,0 +1,1 @@
+export * from './customer-rest.model';
